Simplify process for running SourceCred and improve README
Reduced the 5 commands needed to run a SC instance down to 2 commands. Based on my experience
running an instance and helping others run an instance, the only two things they really want to
do are "load the data" and "run sourcecred". Having separate `site` and `serve` command is just
leaking implementation details to the user, and the graph/score/site commands run fast enough
to group them together into the `yarn start` command. This also limits the chance that the
user updates the weights but forgets to run graph/score before running site/serve and gets
frustrated or confused as to why their cred scores arent updating.

Test Plan: Follow the new instructions in the README for running the instance and ensure that
everything works without errors.